fix(portion): grill fixes — reachable sheets, operable sliders, exact counts
An adversarial review pass found the arithmetic and state layers sound but the
shell around them broken in ways no gate could see. Fixes, worst first:

- Sheets clipped their own action row. showNhamSheet inherited Material's
  9/16 height cap, so a cup portion put Apply at y=594 on a 568pt screen —
  unappliable — and landscape overflowed every vessel by ~104px. The default
  is now isScrollControlled, and a content-hugging body opts into
  NhamSheetSurface(scrollable:) to cap at 90% and scroll past it. This also
  fixes two sheets that were already broken before this branch: the mode
  picker (104px over at large type) and the citations list.

- Both sliders were announced but inoperable. `excludeSemantics: true`
  discarded the Slider's own increase/decrease actions, so a screen reader
  could read "150 g — medium cut" and change nothing. The label now rides the
  wrapper and the actions stay on the slider; the container branch gained
  per-gram divisions so a swipe moves 1 g, not ~56. The previous commit
  message claimed this worked — it did not until now.

- Fractional piece counts truncated. `count` is a plain number server-side and
  the decomposition prompt emits "một lát rưỗi" as 1.5; toInt() made mobile
  build different anchors, claims and committed tiers than web for the same
  meal. count and anchor values are double end-to-end, printed JS-style.

Also: 44pt slider touch target (suppressing the overlay had collapsed it to
16), discrete thumb now lands ON its tick, gram labels scale down instead of
truncating, claimed label no longer clips descenders at 1.3x, kcal preview
matches web to the digit, ticks at the web's 25%, locale changes reach an open
sheet, entrance stagger no longer carries the assumption line, portion applies
skip the confirm debounce as on web.

Hardening, so this class of bug is caught rather than rediscovered:

- test/sheet_overflow_test.dart pumps every content-hugging sheet at 320x568,
  360x640 and landscape, at 1.0x and the 1.3x Dynamic Type ceiling, and fails
  if the primary action can't be reached. Verified non-vacuous.
- test/portion_slider_semantics_test.dart asserts the sliders expose AT
  actions AND that an increase actually moves the portion.
- test/app_fonts.dart loads the real bundled font. flutter_test otherwise
  renders ~1em placeholder glyphs, which invented a clipping bug during review
  that did not exist; any width assertion must call loadAppFonts() first.
- RulerScale moves the grams-to-position mapping out of the widget, where the
  stale-resync trap is now a unit test rather than a latent one.

Files split to stay under the 200-line widget limit: slider shapes, sheet
header, dish line.

+++ apps/mobile-flutter/lib/models/vessel.dart
@@ -36,7 +36,12 @@ sealed class ClientVessel {
 
     if (json['family'] == 'piece') {
       final kind = _byNameOrNull(PieceKind.values, json['kind']);
-      final count = (json['count'] as num?)?.toInt();
+      // NOT toInt(). The server's `count` is a plain number and the
+      // decomposition prompt emits fractions ("nửa" → 0.5, "một lát rưỡi" →
+      // 1.5); `resolvePieceVessel` only rejects count < 1, so 1.5 reaches us.
+      // Truncating it built different anchors from web's for the same meal —
+      // different envelope, different claim, a different committed tier.
+      final count = (json['count'] as num?)?.toDouble();
       if (kind == null || count == null || tier < 1 || tier > 5) return null;
       return PieceVessel(tier: tier, count: count, kind: kind);
     }
@@ -92,7 +97,10 @@ class ContainerVessel extends ClientVessel {
 class PieceVessel extends ClientVessel {
   /// 1–5.
   final int tier;
-  final int count;
+
+  /// How many pieces. Fractional by design — see [ClientVessel.fromJson].
+  final double count;
+
   final PieceKind kind;
 
   const PieceVessel({
